+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Clearing an environment with merged compile requests no longer fails (#2350)
 - Fix "compile_data_json_file" referenced before assignment (#2361)
 - Fix server-autorecompile-wait config option (#2262)
+- Specify the supported values of the 'format' parameter of the OpenAPI endpoint explicitly (#2369)
 
 # Release 2020.4 (2020-09-08)